requirements.txt
streamlit
python-dotenv
openai
MetaTrader5
pandas
requests
